+++ Modules/AssetPipelineEditor/AssetPostprocessors/SketchupMaterialDescriptionPreprocessor.cs
@@ -64,21 +64,21 @@ namespace UnityEditor.AssetImporters
 
             if (description.TryGetProperty("IsTransparent", out floatProperty) && floatProperty == 1.0f)
             {
-                material.SetInt("_Mode", (int)StandardShaderGUI.BlendMode.Transparent);
+                material.SetFloat("_Mode", (float)StandardShaderGUI.BlendMode.Transparent);
                 material.SetOverrideTag("RenderType", "Transparent");
-                material.SetInt("_SrcBlend", (int)UnityEngine.Rendering.BlendMode.One);
-                material.SetInt("_DstBlend", (int)UnityEngine.Rendering.BlendMode.OneMinusSrcAlpha);
-                material.SetInt("_ZWrite", 0);
+                material.SetFloat("_SrcBlend", (float)UnityEngine.Rendering.BlendMode.One);
+                material.SetFloat("_DstBlend", (float)UnityEngine.Rendering.BlendMode.OneMinusSrcAlpha);
+                material.SetFloat("_ZWrite", 0.0f);
                 material.EnableKeyword("_ALPHAPREMULTIPLY_ON");
                 material.renderQueue = (int)UnityEngine.Rendering.RenderQueue.Transparent;
             }
             else
             {
-                material.SetInt("_Mode", (int)StandardShaderGUI.BlendMode.Opaque);
+                material.SetFloat("_Mode", (float)StandardShaderGUI.BlendMode.Opaque);
                 material.SetOverrideTag("RenderType", "");
-                material.SetInt("_SrcBlend", (int)UnityEngine.Rendering.BlendMode.One);
-                material.SetInt("_DstBlend", (int)UnityEngine.Rendering.BlendMode.Zero);
-                material.SetInt("_ZWrite", 1);
+                material.SetFloat("_SrcBlend", (float)UnityEngine.Rendering.BlendMode.One);
+                material.SetFloat("_DstBlend", (float)UnityEngine.Rendering.BlendMode.Zero);
+                material.SetFloat("_ZWrite", 1.0f);
                 material.DisableKeyword("_ALPHATEST_ON");
                 material.DisableKeyword("_ALPHABLEND_ON");
                 material.DisableKeyword("_ALPHAPREMULTIPLY_ON");
